Add files via upload
It contains the classes used for the simulation in two different scripts. Ising_class.py is the script containing the class simulation while Ising_model.pyw is the (no console) script that executes the graphical interface. In the Ising_data folder there are the graphical files used to design the interface. You just have to download this .rar file and execute Ising_model.pyw to use the program.